git-svn-id: https://svn.concord.org/svn/sds/trunk@271 4e27850c-0488-45a1-9bd4-87dd9910adcb

+++ db/migrate/050_create_model_activity_datasets.rb
@@ -4,8 +4,8 @@ class CreateModelActivityDatasets < ActiveRecord::Migration
       t.column :sock_id, :integer
       t.column :created_at, :timestamp
       t.column :name, :string
-      t.column :start_time, :double
-      t.column :end_time, :double
+      t.column :start_time, :float
+      t.column :end_time, :float
     end
     add_index :sds_pas_model_activity_datasets, :sock_id, :name => "sock_id_index"
   end
